When running in a vpc we need public ips
See https://aws.amazon.com/premiumsupport/knowledge-center/create-attach-igw-vpc/, the TLDR for this is that the instances require either a public or elastic ip in order to have comms with the internet, we need these comms for datomic to talk to s3 for transactor logs. It is also possible that this could be setup with a NAT for routing s3 traffic.